+++ src/test/kotlin/one/wabbit/typeclass/plugin/integration/KSerializerBuiltinTest.kt
@@ -1,9 +1,7 @@
 package one.wabbit.typeclass.plugin.integration
 
-import org.junit.Ignore
 import kotlin.test.Test
 
-@Ignore("PHASE11")
 class KSerializerBuiltinTest : IntegrationTestSupport() {
     @Test fun treatsKSerializerAsBuiltinTypeclassWhenFlagEnabled() {
         val source =